exec: Add initial support to IO Redirection
Here document was not implemented yet, also it is currently all under
exec.zig but in the future it can go to its on file.

the parsing of simple command args was also fixed, now it is (hopefully)
correctly parsed the words AND the IO redirection with its io_number.